:sparkles: add env func
