+++ mario_pytorch/play.py
@@ -6,16 +6,10 @@ https://github.com/YuansongFeng/MadMario/blob/master/agent.py
 from pathlib import Path
 
 import torch
-import gym_super_mario_bros
 
-from gym.wrappers import FrameStack
-from nes_py.wrappers import JoypadSpace
-from gym_super_mario_bros.actions import SIMPLE_MOVEMENT
-
-from mario_pytorch.wrappers import SkipFrame, GrayScaleObservation, ResizeObservation
 from mario_pytorch.agent.mario import LearnedMario
-from mario_pytorch.util.get_env_name import get_env_name
 from mario_pytorch.util.config import Config
+from mario_pytorch.util.get_env import get_env
 
 # ----------------------------------------------------------------------
 
@@ -28,15 +22,7 @@ model_name = "mario_net_10.chkpt"
 model_path = checkpoint / model_name
 model = torch.load(model_path)["model"]
 
-# 4 Consecutive GrayScale Frames Set [4, 84, 84]
-env = gym_super_mario_bros.make(
-    get_env_name(config.WORLD, config.STAGE, config.VERSION)
-)
-env = JoypadSpace(env, SIMPLE_MOVEMENT)
-env = SkipFrame(env, skip=config.SKIP_FRAME)
-env = GrayScaleObservation(env)
-env = ResizeObservation(env, shape=config.SHAPE)
-env = FrameStack(env, num_stack=config.NUM_STACK)  # 4Frame まとめて取り出す
+env = get_env(config)
 
 mario = LearnedMario(
     state_dim=(config.NUM_STACK, config.SHAPE, config.SHAPE),
