--- mario_pytorch/util/get_env_name.py
@@ -1,2 +1,0 @@
-def get_env_name(world: int, stage: int, version: int) -> str:
-    return f"SuperMarioBros-{world}-{stage}-v{version}"
